Refactor: genaudit.init_from_command_line_arguments()
