+++ tools/genaudit/genaudit/__init__.py
@@ -3,4 +3,5 @@ from genaudit.repo import GitRepo
 from genaudit.topic import Topic
 from genaudit.render import Renderer
 from genaudit.verify import find_unreferenced_patches, find_misreferenced_pull_request_merges
+from genaudit.base import init_from_command_line_arguments
 from genaudit.refs import *
